chore: Improve Fladder init steps
